feat: Enabling viewing events in GitHub with reorg
With more than 1,000 files in a directory, it's not possible to view all
files and thus requires searching. This reorganization makes it easier
to see all events for a year in the repository by looking in a directory
specifically for that year.

This should not change the public API/schema of /events/events.json,
but it does change how Jekyll represents it internally, thus requiring
the new Liquid filter to manipulate Jekyll's structure into the form
the data used to enter into the templates.

Fixes #1030